Support rvalue parameters in Callback
This unfortunately has the side effect of requiring pass-by value
parameters to be copy-constructed or move-constructed twice instead of
once. This is solvable, but then the side effect is that parameter type
coersion (say from an int to a double or a const char* to a std::string)
is not supported. There seems to be no way to get both that I can find.

std::function makes the same trade off for parameters, and presumably any
type that is passed by value is done so because it is cheap, so this seems
the least-worst solution for now.