Better Callback support for capture-less lambdas
In situations where a capture-less lambda is passed to a function taking a
callback, it would not be accepted without a preceding unary + to force
conversion to a function pointer. While we want the performance
improvement of supporting these as function pointers, we don't want to
require the caller to prefix with unary +.

This makes the type deduction of which initialization function to use
within the Callback constructor, removing this limitation.