Add common parameterized test for FileProtocol
As new FileProtocol derived classes are created, this can cover the common
subset that all FileProtocol classes support, based on the
FileProtocolFlags they define. However, since this is based on the limits
of the FileProtocol interface itself for verification, it is severely
limited for protocols that do not support kInfo and kRead (which nearly
all should do).