Change thread guarantees for Context class
The Context class no longer will hold a mutex while making any call to
external code, including constuctors, destructors, and assignment
operators of contained types. This does mean that client code that
attempts to delete a value on one thread while another is setting
the same value on a different thread could cause corruption for the
value (but not the context itself). This is a reasonable trade off, as
this sort of issue is already present when querying values and then
using the result.