Thread/Job: Give control of fiber threads to caller
FiberJobSystem had an unresolvable race condition in its destructor,
where it could end up deleting a Windows fiber while it was still
running. This change makes it so the Thread handles are returned as well
from CreateFiberThreads, which allows the job system to join all the
threads before deleting the fibers.