Container: Add BufferView classes
BufferView, BufferView2d, and BufferView3d are views onto 1D, 2D, or
3D virtual array of values. The position (origin) for the view can be
set, which will do the minimal number of operations needed to destruct
the values leaving the view and those entering the view. This is done
via a circular buffer which recycles values that leave the view to be
reused for the new values entering the view. As such, a BufferView does
no allocation or deallocation while it is instantiated. Operations can
be performed on the view in virtual space or local (relative) space.