Container: Add OffsetArray family of containers
An OffsetArray is an extension to a one, two, or three dimensional
C-style array, which changes the "zero" index to be at a specific
offset.

For two/three dimensional arrays, only square/cube sized arrays are
supported with the offset affecting all dimensions (aka it is an offset
on the diagonal of the square/cube).